requirements.txt
django-debug-toolbar
django==3.1.1
graphene==2.1.8
graphene-django==2.13.0
django-graphql-auth==0.3.12
django-split-settings==1.0.1
django_versatileimagefield==2.0
django-ckeditor==6.0.0
django-3-jet==1.0.8